Feat/authorization

+++ server/infrastructure/postgres/sql/queries/datasets.sql
@@ -9,12 +9,13 @@ insert into datasets (
     columns,
     alias,
     created_by,
-    updated_by
-) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
+    updated_by,
+    org_id
+) values ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
 returning *;
 
 -- name: GetDataset :one
-select * from datasets where id = $1;
+select * from datasets where id = $1 and org_id = $2;
 
 -- name: UpdateDataset :one
 update datasets
@@ -27,27 +28,34 @@ set
     columns = coalesce($6, columns),
     alias = coalesce($7, alias),
     updated_by = coalesce($8, updated_by)
-where id = $9
+where id = $9 and org_id = $10
 returning *;
 
 -- name: DeleteDataset :exec
-delete from datasets where id = $1;
+delete from datasets where id = $1 and org_id = $2;
 
 -- name: SearchDatasets :many
 select * from datasets
 where 
-    name ilike concat('%', $1, '%') or
-    description ilike concat('%', $1, '%') or
-    alias ilike concat('%', $1, '%')
+    org_id = $1 and
+    (name ilike concat('%', $2, '%') or
+    description ilike concat('%', $2, '%') or
+    alias ilike concat('%', $2, '%'))
 order by 
     case 
-        when alias ilike concat($1, '%') then 1
-        when name ilike concat($1, '%') then 2
-        when name ilike concat('%', $1, '%') then 3
+        when alias ilike concat($2, '%') then 1
+        when name ilike concat($2, '%') then 2
+        when name ilike concat('%', $2, '%') then 3
         else 4
     end,
     created_at desc
-limit $2 offset $3;
+limit $3 offset $4;
 
 -- name: GetDatasetByName :one
-select * from datasets where name = $1;
+select * from datasets where name = $1 and org_id = $2;
+
+-- name: GetDatasetByID :one
+select * from datasets where id = $1;
+
+-- name: ListAllDatasets :many
+select * from datasets;
