+++ server/infrastructure/postgres/sql/queries/projects.sql
@@ -3,9 +3,13 @@ insert into projects (
     name,
     description,
     created_by,
-    updated_by
-) values ($1, $2, $3, $4)
+    updated_by,
+    org_id
+) values ($1, $2, $3, $4, $5)
 returning *;
+
+-- name: ListAllProjects :many
+select * from projects;
 
 -- name: GetProject :one
 select 
@@ -14,7 +18,7 @@ select
     count(pd.dataset_id) as dataset_count
 from projects p
 left join project_datasets pd on p.id = pd.project_id
-where p.id = $1
+where p.id = $1 and p.org_id = $2
 group by p.id;
 
 -- name: UpdateProject :one
@@ -22,12 +26,13 @@ update projects
 set 
     name = coalesce($1, name),
     description = coalesce($2, description),
-    updated_by = coalesce($3, updated_by)
-where id = $4
+    updated_by = coalesce($3, updated_by),
+    org_id = coalesce($4, org_id)
+where id = $5 and org_id = $6
 returning *;
 
 -- name: DeleteProject :exec
-delete from projects where id = $1;
+delete from projects where id = $1 and org_id = $2;
 
 -- name: SearchProjects :many
 SELECT 
@@ -36,17 +41,21 @@ SELECT
 FROM projects p
 LEFT JOIN project_datasets pd ON p.id = pd.project_id
 WHERE 
-    p.name ILIKE concat('%', $1::text, '%') OR
-    p.description ILIKE concat('%', $1::text, '%')
+    p.org_id = $1 AND
+    (p.name ILIKE concat('%', $2::text, '%') OR
+    p.description ILIKE concat('%', $2::text, '%'))
 GROUP BY p.id
 ORDER BY 
     CASE 
-        WHEN p.name ILIKE concat($1::text, '%') THEN 1
-        WHEN p.name ILIKE concat('%', $1::text, '%') THEN 2
+        WHEN p.name ILIKE concat($2::text, '%') THEN 1
+        WHEN p.name ILIKE concat('%', $2::text, '%') THEN 2
         ELSE 3
     END,
     p.created_at DESC
-LIMIT $2 OFFSET $3;
+LIMIT $3 OFFSET $4;
 
 -- name: GetProjectsCount :one
-select count(*) from projects;
+select count(*) from projects where org_id = $1;
+
+-- name: GetProjectByID :one
+select * from projects where id = $1;
